Updated split from libvxf to run independently.

+++ src/core/MemoryStream.cpp
@@ -1,10 +1,33 @@
+////////////////////////////////////////////////////////////////////////////////////////////////////
+// Name:        MemoryStream.cpp
+// Library:     Jameo Core Library
+// Purpose:     In Memory Stream Object
 //
-//  MemoryStream.cpp
-//  jameo
+// Author:      Uwe Runtemund (2015-today)
+// Modified by:
+// Created:     18.08.2015
 //
-//  Created by Uwe Runtemund on 18.08.15.
-//  Copyright (c) 2013 Jameo Software. All rights reserved.
+// Copyright:   (c) 2015 Jameo Software, Germany. https://jameo.de
 //
+// Licence:     The MIT License
+//              Permission is hereby granted, free of charge, to any person obtaining a copy of this
+//              software and associated documentation files (the "Software"), to deal in the
+//              Software without restriction, including without limitation the rights to use, copy,
+//              modify, merge, publish, distribute, sublicense, and/or sell copies of the Software,
+//              and to permit persons to whom the Software is furnished to do so, subject to the
+//              following conditions:
+//
+//              The above copyright notice and this permission notice shall be included in all
+//              copies or substantial portions of the Software.
+//
+//              THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+//              INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+//              PARTICULAR PURPOSE AND NONINFRINGEMENT.IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+//              HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
+//              CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE
+//              OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+////////////////////////////////////////////////////////////////////////////////////////////////////
 
 #include "Precompiled.h"
 
@@ -57,7 +80,7 @@ void MemoryStream::Seek(Integer newPosition)
 void MemoryStream::Move(Integer offset)
 {
 	Integer newPosition = mPosition + offset;
-	mPosition = (newPosition < mStreamlength) ? (uint32)newPosition : mPosition;
+	mPosition = (newPosition < mStreamlength) ? newPosition : mPosition;
 }
 
 Integer MemoryStream::GetPosition()
